+++ NEWS.md
@@ -1,9 +1,7 @@
 # r2dii.utils 0.0.0.9000
 
+* New `create_config()` helps create configuration files for examples and tests.
 * Deprecate `DROPBOX.PATH()`.
-
 * New `path_dropbox_2dii()` improves `DROPBOX.PATH()` with better name and documentation (#2; thanks @cjyetman, @pranavpandya84).
-
 * New functions to handle paths, to handle the configuration file, and miscellaneous.
-
 * Added a `NEWS.md` file to track changes to the package.